Themes: Call startKeyguard() when recreating status bar
mStatusBarKeyguardViewManager holds on to a reference of ScrimController
and we need to update this otherwise the scrim will remain on the screen
and consume all touch events.

This is easily reproduced by changing a theme and then turning the screen
off and on to display the keyguard.  Once the keyguard is dismissed the
scrim remains on the screen.

Change-Id: Ie050a951117ed4aebef4ea3343366b9b9361b8a6